Adds __bridge to some casts of CFSTR().
kAX… strings are generally object-like macros expanding to a CFSTR(…) invocation. CFSTR() itself expands to a builtin when constant CF strings are enabled, and a function call (returning CFStringRef) when disabled.

It’s unclear whether the compiler knows that CFSTR(…) instances are guaranteed to have program lifetime, and therefore that the bridges are irrelevant, or whether it is instead applying some heuristic to the builtin, but for the time being I’ve added the (apparently unnecessary) __bridge to NSString * casts of these constants, for consistency.